Fixed issues with kucoin bot

+++ app/lifeline_crypto_tbot.py
@@ -2,8 +2,6 @@ import asyncio
 
 from aiogram import Dispatcher
 from aiogram import executor
-from aiogram.dispatcher.webhook import get_new_configured_app
-from aiohttp import web
 
 from app import PORT, WEBAPP_HOST, bot
 from app import DEV
@@ -23,56 +21,12 @@ from handler.crypto import send_price_alert
 from handler.crypto import send_trending
 
 
-# from handler.base import error
-# from handler.base import greet
-# from handler.base import start
-# from handler.crypto import coin
-# from handler.crypto import coin_address
-# from handler.crypto import gas
-# from handler.crypto import latest_listings
-# from handler.crypto import priceAlert
-# from handler.crypto import trending
-
-# def main():
-# updater = Updater(token, use_context=True)
-
-# dp = updater.dispatcher
-
-# dp.add_handler(CommandHandler("start", start))
-# dp.add_handler(CommandHandler("help", start))
-# dp.add_handler(CommandHandler("coin", coin))
-# dp.add_handler(CommandHandler("gas", gas))
-# dp.add_handler(CommandHandler("coin_address", coin_address))
-# dp.add_handler(CommandHandler("trending", trending))
-# dp.add_handler(CommandHandler("latest_listings", latest_listings))
-# dp.add_handler(MessageHandler(Filters.status_update.new_chat_members, greet))
-# dp.add_handler(CommandHandler("alert", priceAlert))  # Accessed via /alert
-# # log all errors
-# dp.add_error_handler(error)
-
-# Start the Bot
-#
-# if env == DEV:
-#     updater.start_polling()
-# else:
-#     updater.start_webhook(
-#         listen="0.0.0.0",
-#         port=os.getenv("PORT", "8443"),
-#         url_path=token,
-#         webhook_url=f"{os.getenv('HEROKU_APP_URL')}/{token}",
-#     )
-# # updater.bot.setWebhook(f"{os.getenv('HEROKU_APP_URL')}/{token}")
-
-# Run the bot until you press Ctrl-C or the process receives SIGINT,
-# SIGTERM or SIGABRT. This should be used most of the time, since
-# start_polling() is non-blocking and will stop the bot gracefully.
 async def on_startup(dp: Dispatcher):
-    await bot.delete_webhook()
-    await bot.set_webhook(WEBHOOK_URL)
+    if ENV != DEV:
+        await bot.delete_webhook()
+        await bot.set_webhook(WEBHOOK_URL)
     setup_handlers(dp)
     asyncio.create_task(kucoin_bot())
-    # # await send_message(channel_id=TELEGRAM_CHAT_ID, text="This is a test")
-    # await task
 
 
 def setup_handlers(dp: Dispatcher):
